kevin

+++ pa2/src/dr/drAstar.cpp
@@ -399,7 +399,7 @@ EPathDir DrAstar::pathDir(const Point3D& p1, const Point3D& p2) const {
     } 
     else assert(false);
 }
-float DrAstar::costG(AstarNode* u, AstarNode* v, int groupId)  {
+int DrAstar::costG(AstarNode* u, AstarNode* v, int groupId)  {
     return abs(u->coor[0] - v->coor[0]) + abs(u->coor[1] - v->coor[1]) + (abs(u->coor.layer() - v->coor.layer()) * _astarParam.viaCost)
         + _db.capacityTable().cost(u->coor, v->coor);
 }
